Make alc and alcc work with files >4GB by Skolnick, progress bar fix still missing

+++ trunk/src/utils/aLinkCreator/src/ed2khash.h
@@ -48,7 +48,7 @@ class Ed2kHash:public MD4
 
     wxArrayString m_ed2kArrayOfHashes;
     wxString m_filename;
-    size_t m_fileSize;
+    uint64_t m_fileSize;
 
   protected:
